fix:工时比例改为小数

+++ mes-modules/system/src/main/java/com/richfit/mes/sys/entity/dto/TenantUpdateUserDto.java
@@ -59,5 +59,5 @@ public class TenantUpdateUserDto extends BaseDto<TenantUser> {
     private String userRoleType;
 
     @ApiModelProperty(value = "工时比例")
-    private Integer ratioHours;
+    private Double ratioHours;
 }
